Resolver: Simple resolving
This is still a WIP but the main idea is working.
I've decided to write the resolved data directly into a result
json string as this is very quick.
Note that the allocation can probably still be quicker if we use some
sort of buffer but that's something to inspect later.